move tests that require node from test/node to node-test
Is a pain to include all sub dirs in test apart from node
