+++ node-test/function/alegbra/sparse/cs_amd.test.js
@@ -1,8 +1,8 @@
 var assert = require('assert');
-var approx = require('../../../../../tools/approx');
-var market = require('../../../../../tools/matrixmarket');
-var math = require('../../../../../index').create();
-math.import(require('../../../../../lib/function/algebra/sparse/cs_amd'));
+var approx = require('../../../../tools/approx');
+var market = require('../../../../tools/matrixmarket');
+var math = require('../../../../index').create();
+math.import(require('../../../../lib/function/algebra/sparse/cs_amd'));
 
 var cs_amd = math.sparse.cs_amd;
