+++ node-test/function/alegbra/sparse/cs_lu.test.js
@@ -1,8 +1,18 @@
-var approx = require('../../../../../tools/approx'),
-  math = require('../../../../../index'),
-  market = require('../../../../../tools/matrixmarket');
+var assert = require('assert');
+var approx = require('../../../../tools/approx');
+var market = require('../../../../tools/matrixmarket');
+var math = require('../../../../index').create();
 
-describe('slu - matrix market', function () {
+math.import(require('../../../../lib/function/algebra/sparse/cs_permute'));
+math.import(require('../../../../lib/function/algebra/sparse/cs_lu'));
+math.import(require('../../../../lib/function/algebra/sparse/cs_sqr'));
+
+var cs_permute = math.sparse.cs_permute;
+var cs_lu = math.sparse.cs_lu;
+var cs_sqr = math.sparse.cs_sqr;
+
+describe('cs_lu', function () {
+
   it('should decompose matrix, 48 x 48, natural ordering (order=0), full pivoting, matrix market', function (done) {
     // import matrix
     market.import('tools/matrices/bcsstk01.tar.gz', ['bcsstk01/bcsstk01.mtx'])
@@ -10,11 +20,14 @@ describe('slu - matrix market', function () {
         // matrix
         var m = matrices[0];
 
-        // full pivoting
-        var r = math.slu(m, 0, 0.001);
+        // symbolic ordering and analysis, order = 0
+        var s = cs_sqr(0, m, false);
 
-        // verify M[p,q]=L*U
-        approx.deepEqual(_permute(m, r.p, r.q).valueOf(), math.multiply(r.L, r.U).valueOf());
+        // full pivoting
+        var r = cs_lu(m, s, 0.001);
+
+        // verify
+        approx.deepEqual(cs_permute(m, r.pinv, s.q, true).valueOf(), math.multiply(r.L, r.U).valueOf());
 
         // indicate test has completed
         done();
@@ -32,11 +45,14 @@ describe('slu - matrix market', function () {
         // matrix
         var m = matrices[0];
 
-        // full pivoting
-        var r = math.slu(m, 1, 0.001);
+        // symbolic ordering and analysis, order = 1
+        var s = cs_sqr(1, m, false);
 
-        // verify M[p,q]=L*U
-        approx.deepEqual(_permute(m, r.p, r.q).valueOf(), math.multiply(r.L, r.U).valueOf());
+        // full pivoting
+        var r = cs_lu(m, s, 0.001);
+
+        // verify
+        approx.deepEqual(cs_permute(m, r.pinv, s.q, true).valueOf(), math.multiply(r.L, r.U).valueOf());
 
         // indicate test has completed
         done();
@@ -54,11 +70,14 @@ describe('slu - matrix market', function () {
         // matrix
         var m = matrices[0];
 
-        // full pivoting
-        var r = math.slu(m, 2, 0.001);
+        // symbolic ordering and analysis, order = 2
+        var s = cs_sqr(2, m, false);
 
-        // verify M[p,q]=L*U
-        approx.deepEqual(_permute(m, r.p, r.q).valueOf(), math.multiply(r.L, r.U).valueOf());
+        // full pivoting
+        var r = cs_lu(m, s, 0.001);
+
+        // verify
+        approx.deepEqual(cs_permute(m, r.pinv, s.q, true).valueOf(), math.multiply(r.L, r.U).valueOf());
 
         // indicate test has completed
         done();
@@ -76,11 +95,14 @@ describe('slu - matrix market', function () {
         // matrix
         var m = matrices[0];
 
-        // full pivoting
-        var r = math.slu(m, 3, 0.001);
+        // symbolic ordering and analysis, order = 3
+        var s = cs_sqr(3, m, false);
 
-        // verify M[p,q]=L*U
-        approx.deepEqual(_permute(m, r.p, r.q).valueOf(), math.multiply(r.L, r.U).valueOf());
+        // full pivoting
+        var r = cs_lu(m, s, 0.001);
+
+        // verify
+        approx.deepEqual(cs_permute(m, r.pinv, s.q, true).valueOf(), math.multiply(r.L, r.U).valueOf());
 
         // indicate test has completed
         done();
@@ -90,41 +112,4 @@ describe('slu - matrix market', function () {
         done(error);
       });
   });
-
-  /**
-   * C = A(p,q) where p is the row permutation vector and q the column permutation vector.
-   */
-  var _permute = function (A, pinv, q) {
-    // matrix arrays
-    var values = A._values;
-    var index = A._index;
-    var ptr = A._ptr;
-    var size = A._size;
-    // columns
-    var n = size[1];
-    // c arrays
-    var cvalues = [];
-    var cindex = [];
-    var cptr = [];
-    // loop columns
-    for (var k = 0 ; k < n ; k++) {
-      cptr[k] = cindex.length;
-      // column in C
-      var j = q ? (q[k]) : k;
-      // values in column j
-      for (var t = ptr[j]; t < ptr[j + 1]; t++) {
-        cvalues.push(values[t]);
-        cindex.push(pinv ? (pinv[index[t]]) : index[t]);
-      }
-    }
-    cptr[n] = cindex.length;
-    // return matrix
-    return new math.type.SparseMatrix({
-      values: cvalues,
-      index: cindex,
-      ptr: cptr,
-      size: size,
-      datatype: A._datatype
-    });
-  };
 });
