+++ node-test/function/arithmetic/multiply.test.js
@@ -1,8 +1,8 @@
 // test multiply
 var assert = require('assert'),
-  math = require('../../../../index'),
-  approx = require('../../../../tools/approx'),
-  market = require('../../../../tools/matrixmarket'),
+  math = require('../../../index'),
+  approx = require('../../../tools/approx'),
+  market = require('../../../tools/matrixmarket'),
   multiply = math.multiply,
   divide = math.divide,
   matrix = math.matrix,
